QR_Attendance_System/requirements.txt
Flask
python-dotenv
qrcode
smtplib
email
Werkzeug
opencv-python-headless
gspread
oauth2client
Pillow  # Required by the `qrcode` library for image processing